Separate geometry from BC
