+++ src/CabanaPD.hpp
@@ -20,6 +20,7 @@
 #include <CabanaPD_Fields.hpp>
 #include <CabanaPD_Force.hpp>
 #include <CabanaPD_ForceModels.hpp>
+#include <CabanaPD_Geometry.hpp>
 #include <CabanaPD_Input.hpp>
 #include <CabanaPD_Integrate.hpp>
 #include <CabanaPD_Output.hpp>
