+++ src/CabanaPD_Boundary.hpp
@@ -16,217 +16,11 @@
 
 #include <Cabana_Core.hpp>
 
+#include <CabanaPD_Geometry.hpp>
 #include <CabanaPD_Timer.hpp>
 
 namespace CabanaPD
 {
-
-struct RectangularPrism
-{
-};
-struct Cylinder
-{
-};
-struct Custom
-{
-};
-
-// User-specifed custom boundary. Must use the signature:
-//    bool operator()(PositionType&, const int)
-template <typename UserFunctor>
-struct RegionBoundary
-{
-    UserFunctor _user_functor;
-
-    RegionBoundary( UserFunctor user )
-        : _user_functor( user )
-    {
-    }
-
-    template <class PositionType>
-    KOKKOS_INLINE_FUNCTION bool inside( const PositionType& x,
-                                        const int pid ) const
-    {
-        return user( x, pid );
-    }
-};
-
-// Define a subset of the system as the boundary with a rectangular prism.
-template <>
-struct RegionBoundary<RectangularPrism>
-{
-    double low_x;
-    double high_x;
-    double low_y;
-    double high_y;
-    double low_z;
-    double high_z;
-
-    RegionBoundary( const double _low_x, const double _high_x,
-                    const double _low_y, const double _high_y,
-                    const double _low_z, const double _high_z )
-        : low_x( _low_x )
-        , high_x( _high_x )
-        , low_y( _low_y )
-        , high_y( _high_y )
-        , low_z( _low_z )
-        , high_z( _high_z ){};
-
-    template <class PositionType>
-    KOKKOS_INLINE_FUNCTION bool inside( const PositionType& x,
-                                        const int pid ) const
-    {
-        return ( x( pid, 0 ) >= low_x && x( pid, 0 ) <= high_x &&
-                 x( pid, 1 ) >= low_y && x( pid, 1 ) <= high_y &&
-                 x( pid, 2 ) >= low_z && x( pid, 2 ) <= high_z );
-    }
-};
-
-// Define a subset of the system as the boundary with a cylinder.
-template <>
-struct RegionBoundary<Cylinder>
-{
-    double radius_in;
-    double radius_out;
-    double low_z;
-    double high_z;
-    double x_center;
-    double y_center;
-
-    RegionBoundary( const double _radius_in, const double _radius_out,
-                    const double _low_z, const double _high_z,
-                    double _x_center = 0.0, double _y_center = 0.0 )
-        : radius_in( _radius_in )
-        , radius_out( _radius_out )
-        , low_z( _low_z )
-        , high_z( _high_z )
-        , x_center( _x_center )
-        , y_center( _y_center ){};
-
-    template <class PositionType>
-    KOKKOS_INLINE_FUNCTION bool inside( const PositionType& x,
-                                        const int pid ) const
-    {
-        double rsq = ( x( pid, 0 ) - x_center ) * ( x( pid, 0 ) - x_center ) +
-                     ( x( pid, 1 ) - y_center ) * ( x( pid, 1 ) - y_center );
-        return ( rsq >= radius_in * radius_in &&
-                 rsq <= radius_out * radius_out && x( pid, 2 ) >= low_z &&
-                 x( pid, 2 ) <= high_z );
-    }
-};
-
-template <class MemorySpace, class BoundaryType>
-struct BoundaryIndexSpace;
-
-// FIXME: fails for some cases if initial guess is not sufficient.
-template <class MemorySpace, class GeometryType>
-struct BoundaryIndexSpace<MemorySpace, RegionBoundary<GeometryType>>
-{
-    using index_view_type = Kokkos::View<std::size_t*, MemorySpace>;
-    index_view_type _view;
-    index_view_type _count;
-
-    Timer _timer;
-
-    // Default for empty case.
-    BoundaryIndexSpace() {}
-
-    // Construct from region (search for boundary particles).
-    template <class ExecSpace, class Particles>
-    BoundaryIndexSpace( ExecSpace exec_space, Particles particles,
-                        std::vector<RegionBoundary<GeometryType>> planes,
-                        const double initial_guess )
-    {
-        _timer.start();
-
-        _view = index_view_type( "boundary_indices",
-                                 particles.localOffset() * initial_guess );
-        _count = index_view_type( "count", 1 );
-
-        for ( RegionBoundary plane : planes )
-        {
-            update( exec_space, particles, plane );
-        }
-        // Resize after all planes searched.
-        auto count_host =
-            Kokkos::create_mirror_view_and_copy( Kokkos::HostSpace{}, _count );
-        if ( count_host( 0 ) < _view.size() )
-        {
-            Kokkos::resize( _view, count_host( 0 ) );
-        }
-
-        _timer.stop();
-    }
-
-    template <class ExecSpace, class Particles>
-    void update( ExecSpace, Particles particles,
-                 RegionBoundary<GeometryType> region )
-    {
-        auto count_host =
-            Kokkos::create_mirror_view_and_copy( Kokkos::HostSpace{}, _count );
-        auto init_count = count_host( 0 );
-
-        auto index_space = _view;
-        auto count = _count;
-        auto x = particles.sliceReferencePosition();
-        // TODO: configure including frozen particles.
-        Kokkos::RangePolicy<ExecSpace> policy( 0, particles.localOffset() );
-        auto index_functor = KOKKOS_LAMBDA( const std::size_t pid )
-        {
-            if ( region.inside( x, pid ) )
-            {
-                // Resize after count if needed.
-                auto c = Kokkos::atomic_fetch_add( &count( 0 ), 1 );
-                if ( c < index_space.size() )
-                {
-                    index_space( c ) = pid;
-                }
-            }
-        };
-
-        Kokkos::parallel_for( "CabanaPD::BC::update", policy, index_functor );
-        Kokkos::deep_copy( count_host, _count );
-        if ( count_host( 0 ) > index_space.size() )
-        {
-            Kokkos::resize( index_space, count_host( 0 ) );
-            Kokkos::deep_copy( count, init_count );
-            Kokkos::parallel_for( "CabanaPD::BC::update", policy,
-                                  index_functor );
-        }
-    }
-
-    auto time() { return _timer.time(); };
-};
-
-template <class MemorySpace>
-struct BoundaryIndexSpace<MemorySpace, Custom>
-{
-    using index_view_type = Kokkos::View<std::size_t*, MemorySpace>;
-    index_view_type _view;
-
-    // Construct from a View of boundary particles.
-    BoundaryIndexSpace( index_view_type input_view )
-        : _view( input_view )
-    {
-    }
-};
-
-template <class ExecSpace, class Particles, class BoundaryType>
-auto createBoundaryIndexSpace( ExecSpace exec_space, Particles particles,
-                               std::vector<BoundaryType> planes,
-                               const double initial_guess )
-{
-    using memory_space = typename Particles::memory_space;
-    return BoundaryIndexSpace<memory_space, BoundaryType>(
-        exec_space, particles, planes, initial_guess );
-}
-
-template <class BoundaryParticles>
-auto createBoundaryIndexSpace( BoundaryParticles particles )
-{
-    using memory_space = typename BoundaryParticles::memory_space;
-    return BoundaryIndexSpace<memory_space, Custom>( particles );
-}
 
 struct ForceValueBCTag
 {
